backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
pydantic==2.5.0
pydantic-settings==2.1.0
python-dotenv==1.0.0
httpx==0.25.2
email-validator>=2.0.0

# MongoDB
motor==3.3.2
pymongo==4.6.1

# Optional: SQL Database (can be removed if using only MongoDB)
sqlalchemy==2.0.23
# psycopg2-binary==2.9.9  # Uncomment for PostgreSQL
# alembic==1.12.1  # Uncomment for migrations

# AI/ML (optional - install as needed)
# openai==1.3.7
google-generativeai>=0.3.2
# sentence-transformers==2.2.2
# faiss-cpu==1.7.4

# Face Recognition & Computer Vision
opencv-python==4.8.1.78
face-recognition==1.3.0
dlib==19.24.2
numpy==1.24.3
Pillow==10.1.0
deepface==0.0.79
mediapipe==0.10.7

# Utilities
PyPDF2==3.0.1

# Testing
pytest==7.4.3
pytest-asyncio==0.21.1
httpx==0.25.2
pytest-cov==4.1.0

# Background tasks (optional)
# celery==5.3.4
# redis==5.0.1
